chore(benchmark): renumber S19-msvc-flock-trap to S20
#191 also claims S19 (S19-phantom-update-stale-host, opened 2026-09-08).
Yielding the slot keeps both PRs mergeable without a duplicate-S19 conflict.
Registry counts stay at 57 tasks; the scoring total stays 5700.
