+++ benchmark/README.md
@@ -55,7 +55,7 @@ honestly instead of quietly fixing it and pretending nothing happened).
 | S15-slot-error-boundary-crash | Static | After a feature release the pending-attachment dock vanishes entirely: a dangling `busy` identifier (another component's state) throws only when a chip renders and the phase guard short-circuits open — can it find the pre-existing line behind the misleading new diff and design a data-present render regression |
 | S16-self-host-upgrade-trap | Static | An agent runs the global dsh upgrade from inside its own session on the running host: npm replaces the live package tree, the GUI dies mid-call, and the interrupted install leaves the package present but the `dsh` command gone — can it diagnose the self-upgrade failure, repair via an external pinned re-install, and state the hand-off protocol (host stopped BEFORE npm so nothing crashes mid-install) |
 | S17-external-ui-plugin-onboarding-trap | Static | A hand-written external web UI plugin is inserted into a running profile: one raw-ESM client bundle fails the whole combo so ZERO plugins register (the error names an innocent first-awaited entry), the repackaged plugin then hits a cross-entry slot declaration race, and every edit needs a full host restart (Windows tree-kill or EADDRINUSE) - diagnose all three from the evidence pack |
-| S19-msvc-flock-trap | Static | Upgrading to 0.1.3-alpha.1 on a Windows machine without MSVC: `pnpm install` dies on the fs-ext native build even though Windows never calls `flock` (named kernel semaphore) — can it diagnose the static-import trap (`--ignore-scripts` cannot skip it) and plan the pnpm-patch fix instead of demanding Visual Studio |
+| S20-msvc-flock-trap | Static | Upgrading to 0.1.3-alpha.1 on a Windows machine without MSVC: `pnpm install` dies on the fs-ext native build even though Windows never calls `flock` (named kernel semaphore) — can it diagnose the static-import trap (`--ignore-scripts` cannot skip it) and plan the pnpm-patch fix instead of demanding Visual Studio |
 | M2-optional-dep-trap | Hands-on | The plugin declares an optional dependency but imports it unconditionally at top level (the comment says optional is harmless): does it fix the dependency contract instead of wrapping the import, and prove it with a cold boot |
 | M3-session-projection | Hands-on | A self-assembled profile mounts dsh-tool-todo without the sessionProjections service: does it fix the composition (never edit shipped packages) so the tree activates while the todo tool survives in the final composition |
 | M4-peer-prerelease-range | Hands-on | A peer lower bound written as ^0.1.0-rc.8 does not match 0.1.2-alpha.2 under npm semver's prerelease rule: does it rewrite the bound to the target cohort instead of widening it into a meaningless range |
