+++ benchmark/docs/scoring.md
@@ -50,7 +50,7 @@ conditions with the same pilot snapshot before reporting new scores.
 | S15-slot-error-boundary-crash | skills/plugin-runtime-debug (silent slot-unmount family; same 2026-09-03 session's dangling-busy crash) | Five × 20: free identifier + || short-circuit + chip-present latency / error-boundary whole-entry unmount (console-only) / attribution vs the newest diff (rollback or minimal render mount) / fix + hardening (?? [], closest?.()) / data-present render smoke + no-undef |
 | S16-self-host-upgrade-trap | skills/plugin-upgrade (global host-upgrade agent discipline; real 2026-09-03 dsh alpha.5→rc.1 incident) | Five × 20: self-upgrade structural root cause (session IS the host, npm replaces the live tree, call never returns) / interrupted-install signature (shims missing, content ≠ working install) / external pinned re-install repair / protocol: refuse + hand off, host stopped BEFORE npm so nothing crashes mid-install / prevention guard + post-upgrade checklist |
 | S17-external-ui-plugin-onboarding-trap | skills/plugin-runtime-debug (external-plugin onboarding symptom families; real 2026-09-04 dsh-v0.1.3-alpha.1 source-host incident) | Five x 20: combo total-failure root cause (one raw-ESM bundle fails the whole classic-script combo, zero registrations, error names the innocent first-awaited entry) / diagnosis discipline (bisect inserts, classic-script parse with vm.Script, __ModuleLoader__.load + require(react) + inject/apply) / cross-entry slot declaration (ctx.slots.inject wrapper, registrant field whitelist) / dev-loop discipline (boot-assembled combo, full restart, Windows tree-kill vs EADDRINUSE) / prevention (startup scan names the culprit, authoring template) |
-| S19-msvc-flock-trap | DSH-0.1.3-A1-03 · Windows install fails on the `fs-ext` native build; the runtime never calls `flock` on Windows; rollup R-08 install-channel pitfalls | Report checkpoints: root cause fs-ext + node-gyp/MSVC (20) + static-import trap — `--ignore-scripts` cannot skip a statically imported module (20) + Windows lock path (named kernel semaphore, flock never called; partial credit 12) (20) + pnpm-patch recipe (win32 skips the native build, pure-JS fallback, Linux/macOS keep native; incomplete recipe 15) (25) + no-Visual-Studio discipline (10) + card citation DSH-0.1.3-A1-03 (5). A plan relying on installing Visual Studio ⇒ capped at 50; relying on `--ignore-scripts` alone ⇒ capped at 40; fixture modified ⇒ flat 0 |
+| S20-msvc-flock-trap | DSH-0.1.3-A1-03 · Windows install fails on the `fs-ext` native build; the runtime never calls `flock` on Windows; rollup R-08 install-channel pitfalls | Report checkpoints: root cause fs-ext + node-gyp/MSVC (20) + static-import trap — `--ignore-scripts` cannot skip a statically imported module (20) + Windows lock path (named kernel semaphore, flock never called; partial credit 12) (20) + pnpm-patch recipe (win32 skips the native build, pure-JS fallback, Linux/macOS keep native; incomplete recipe 15) (25) + no-Visual-Studio discipline (10) + card citation DSH-0.1.3-A1-03 (5). A plan relying on installing Visual Studio ⇒ capped at 50; relying on `--ignore-scripts` alone ⇒ capped at 40; fixture modified ⇒ flat 0 |
 | M6-sleep-tool | DSH-0.1.2-A2-03 peer hygiene (bare cordis contract, ContentBlock → dsh-llm); real @huanlin/dsh-plugin-sleep e25a4a9 | Static 50 (bare cordis gone 12 + @deepseek-ai/cordis 10 + dsh-tools floor 12 + dsh-llm peer 10 + meta optional 6) + headless activation 25 + diagnose 15 (exists 5, names 5, A2-03 3, R-01 2) + release 10; memo bait (keep bare cordis) caps at 60, changed-but-unfixed caps at 40 (M4 precedent) |
 | M7-d399-overlay | DSH-0.1.2-A1-25 client-runtime removal; DSH-0.1.2-A1-19 roster anchor | Static 50 (runtimeGone 10 + inject recomposed 10 + ISessions annotation 10 + cohort 12 + scoped cordis 8) + deploy 25 + diagnose 15 (exists 5, names 5, A1-25 3, R-01 2) + release 10; runtime retained caps at 20 |
 | M8-brand-text | DSH-0.1.2-A1-25 (store engine → dsh-client-store); rollup R-06 baseline attribution (H2 precedent) | Static 50 (runtimeGone + storeTypeSource + inject recomposed + cordis ^4.0.1 without -rc + cohort, 10 each) + deploy 25 + diagnose 15 incl. pre-existing-red-test honesty item + release 10; runtime retained caps at 20, store-bait retained caps at 60 |
