+++ benchmark/tasks/S20-msvc-flock-trap/environment/Dockerfile
