+++ benchmark/tasks/S20-msvc-flock-trap/environment/fixture/manifest-excerpt.json
