+++ benchmark/tasks/S20-msvc-flock-trap/environment/fixture/win32-excerpt.ts
